Split lib/notifications/server into separate actions
