+++ pages/api/send-manage-link.ts
@@ -1,7 +1,8 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { constants } from "http2";
-import { loadSubscription } from "lib/notifications/server";
 import sendManageEmail from "lib/notifications/emails/send-manage-email";
+import { loadSubscription } from "lib/notifications/actions/loadSubscription";
+import loadConfig from "lib/notifications/helpers/loadConfig";
 
 type FormData = {
     emailAddress: string;
@@ -14,6 +15,8 @@ const isFormData = (data: unknown): data is FormData => {
 }
 
 const SendManageLink = async (req: NextApiRequest, res: NextApiResponse) => {
+    const config = loadConfig();
+
     if (req.method !== 'POST') {
         return res
             .status(constants.HTTP_STATUS_METHOD_NOT_ALLOWED)
@@ -27,7 +30,7 @@ const SendManageLink = async (req: NextApiRequest, res: NextApiResponse) => {
             .json({ message: 'Incomplete alert form' });
     }
 
-    const subscription = await loadSubscription(formData.emailAddress);
+    const subscription = await loadSubscription(config, formData.emailAddress);
 
     if (subscription) {
         await sendManageEmail(subscription)
